+++ src/main/java/cn/edu/xmu/whiteboard/controller/dto/pb/Roundness.java
@@ -3,12 +3,12 @@ package cn.edu.xmu.whiteboard.controller.dto.pb;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class Roundness {
-    private long type;
+    private double type;
 
     public Roundness() {}
 
     @JsonProperty("type")
-    public long getType() { return type; }
+    public double getType() { return type; }
     @JsonProperty("type")
-    public void setType(long value) { this.type = value; }
+    public void setType(double value) { this.type = value; }
 }
